Added rho_l and rho_r option
Added the option to have a separate rho value to the left and right of the cutoff. This option was necessary because if you specify the bandwidth with the h() option rdrobust by default sets b_l equal to h_l and b_r equal to h_r. Rho is therefore automatically set to 1 (unless the rho option is specified) on both sides. It was possible to set the same rho value for both sides using the rho() option but it was not possible to set a rho to the left of the cutoff and a rho to the right of the cutoff if manually selecting the bandwidth. Kept the default rho value of 0.