Returned Mata results directly to r()
Returned the Mata vectors of standard errors and lower/upper confidence interval limits directly to r() in Stata instead of creating a Stata matrix equal to the returned results from the Mata function and returning the Stata matrix to r(). Added a "return add" command to post the current r() values to return().