Added t statistic to e(table)
Added a t statistic based off the p value. Set t statistic to 0 if p==1. Set the t statistic to minfloat() if p==0 and the coefficient estimate is less than 0. Set the t statistic to maxfloat() if p==0 and the coefficient estimate is greater than 0. Otherwise calculated the t statistic using invttail based off the p value, residual degrees of freedom, and the sign of the coefficient estimate.