Added r(table) values and sample sizes
Created a matrix with the r(table) values: the coefficient estimate, the standard error, the z statistic, the p value, and the lower/upper levels of the confidence interval. Marked the values as missing if the standard errors could not be calculated. Exited with error it neither the coefficients nor the standard errors could be estimated. Returned the sample size and subpopulation size as matrices.